ptune: Release v1.6.0
Still not v2 yet... but we're getting closer! This build is intended to maintain compatibility with stock public A16 builds from Google. It also employs a number of other changes, upon which further improvements to overall system stability and other minor adjustments have been made to enhance the user experience.

install:
- Support Android 16 initial release (public stable)
- Continue to support Android 15 (post-November 2024) until further notice
- Support Pixel 9a
- Support Pixel 10 series
- Support Pixel 11 series
- Support Pixel 12 series (with no known codename for 12a, alas `slowltexx` as temp)
- Stop symlinking service script into Magisk service.d directory, it actually runs fine now from only the module's directory (possibly an old bug in Magisk?)

init:
- Introduce `task_profiles ProcessCapacityMax MaxPerformance` to all patched services
- Support 32-bit Zygote on 64-bit systems when present (still no support for 32-bit only systems)
- Support HWC 2.4 for legacy Tensor
- Support Google `radioext`
- Add missing `cameraserver` patches
- Add missing bootloop protection for `keystore2`

service:
- Support adjusting size of ZRAM block device `zram0`, now changed to be total physical RAM - 1GB (i.e. Pixel 6 Pro with 12GB will now have 11GB of zram, effectively creating 23GB of RAM under lz44eh's max compression ratio of 2x, your mileage may vary)
- Drop Pixel scheduler frequency back down to 120Hz
- Allow ZRAM to reach 99% before triggering LMKD (Low Memory Killer Daemon)
- Enable support for transparent huge pages (THP) in the kernel
- Drop swappiness from 1 to 0
- Drop VFS cache pressure from 5 to 1
- Tell SurfaceFlinger to sample regions at 120Hz instead of 240Hz like before
- Adjust all schedgroups to max out their uclamps
- Adjust all cpusets (except for camera sets) to run on all cores
- Read from `/dev/cpuset/cpus` to obtain max core range for cpusets running on all cores (fixes the idle 9th core on P8 series)
- Fix `camera-daemon` cpuset to use all cores (like stock) instead of only little
- Log all ptune actions under `/cache/ptune.log` (alternatively `/data/cache/ptune.log` if your device does not bind the cache folder anymore)
- Handle execution flow all from the service context, no need for a post-fs-data script

+++ module/service.sh
@@ -1,9 +1,12 @@
 #!/system/bin/sh
 
+LOGFILE="/cache/ptune.log"
+
 #MODDIR=${0%/*}
-SHDIR="$(dirname $0)"
-MAGISKSD=/dev/adb/service.d
-MAGISKPFDD=/dev/adb/post-fs-data.d
+DIRSH="$(dirname $0)"
+DIRBIN="$DIRSH/bin"
+
+PATH="$DIRBIN:$PATH"
 
 VM=/proc/sys/vm
 KR=/proc/sys/kernel
@@ -11,6 +14,8 @@ CS=/dev/cpuset
 VS=/proc/vendor_sched
 VSG=$VS/groups
 FREQ=/sys/devices/system/cpu/cpufreq
+MM=/sys/kernel/mm
+THP=$MM/transparent_hugepage
 
 F0=$FREQ/policy0
 F1=$FREQ/policy4
@@ -88,7 +93,19 @@ bootcomplete() {
 
 #########
 
+logwipe() {
+  rm -f "$LOGFILE"
+}
+
+log() {
+  ls="($DIRSH) ptune: $1"
+  echo "$ls" >> "$LOGFILE"
+  echo "$ls"
+}
+
 ptune() {
+
+cpuf="$(cat $CS/cpus)"
 
 # Allow vendor scheduler groups to fully utilize cores
 echo "2048 2048 2048 2048 2048 2048 2048 2048" > $VS/util_threshold
@@ -104,40 +121,41 @@ echo "2048 2048 2048 2048 2048 2048 2048 2048" > $VS/util_threshold
 # prefer big
 # prefer idle
 # task spread
-schedgroup bg        0     158 0 1 1 #0 158
-schedgroup cam       0    1024 0 1 0 #0 1024 (?¿?)
-schedgroup cam_power 0    1024 0 1 0 #0 1024 (?¿?)
-schedgroup dex2oat   0    1024 0 1 0 #0 1024 (?¿?)
-schedgroup fg        0    1024 0 1 0 #0 490
-schedgroup fg_wi     0     490 0 1 0 #0 757 (?¿?)
-schedgroup nnapi     0    1024 0 1 0 #0 1024 (?¿?)
-schedgroup ota       0    1024 0 1 0 #0 1024 (?¿?)
-schedgroup rt        0     158 0 1 1 #0 158
-schedgroup sf        0    1024 0 1 0 #0 490 (?¿?)
-schedgroup sys       0    1024 0 1 0 #0 490 (?¿?)
-schedgroup sys_bg    0     490 0 1 0 #0 158
-schedgroup ta        0    1024 0 1 0 #0 1024
+schedgroup bg        0    1024 0 1 1 #0     512
+schedgroup cam       0    1024 0 1 0 #1    1024
+schedgroup cam_power 0    1024 0 1 0 #0    1024
+schedgroup dex2oat   0    1024 0 1 0 #0     615
+schedgroup fg        0    1024 0 1 0 #0    1024
+schedgroup fg_wi     0    1024 0 1 0 #0    1024
+schedgroup nnapi     0    1024 0 1 0 #225  1024
+schedgroup ota       0    1024 0 1 0 #0     512
+schedgroup rt        0    1024 0 1 0 #0    1024
+schedgroup sf        0    1024 0 1 0 #0    1024
+schedgroup sys       0    1024 0 1 0 #0    1024
+schedgroup sys_bg    0    1024 0 1 0 #0     512
+schedgroup ta        0    1024 0 1 0 #1    1024
 
 # cpuset | cpus
-cpuset background                   0-3 #0-3
-cpuset camera-daemon                0-7 #0-7 (?¿?)
-cpuset camera-daemon-high-group     0-7 #0-7 (?¿?)
-cpuset camera-daemon-mid-group      0-7 #0-7 (?¿?)
-cpuset camera-daemon-mid-high-group 0-7 #0-7 (?¿?)
-cpuset foreground                   0-7 #0-5
-cpuset foreground_window            0-5 #0-6
-cpuset restricted                   0-3 #0-3
-cpuset system                       0-7 #0-7 (custom, !sys)
-cpuset system-background            0-5 #0-3
-cpuset top-app                      0-7 #0-7
+cpuset background                   "$cpuf" #0-3
+cpuset camera-daemon                "$cpuf" #0-7
+cpuset camera-daemon-high-group     6-7     #6-7
+cpuset camera-daemon-mid-group      4-5     #4-5
+cpuset camera-daemon-mid-high-group 4-7     #4-7
+cpuset foreground                   "$cpuf" #0-5
+cpuset foreground_window            "$cpuf" #0-5
+cpuset restricted                   0-3     #0-3
+cpuset system                       "$cpuf" #0-7 (custom, !sys)
+cpuset system-background            "$cpuf" #0-3
+cpuset top-app                      "$cpuf" #0-7
 
 # Give our CPU a lunch break when it wants one
+# down delay | up delay
 delayfreqs 0 0 #5000 0
 
 # adpf rampup multiplier
 # latency in nanoseconds
 # reduce prefer idle
-sched 1 4166666 0 #2 8000000 0
+sched 1 8333333 0 #2 8000000 0
 
 # Speed up disk access
 # async depth
@@ -149,47 +167,74 @@ blocksched sdd mq-deadline 100 16384 #62 62
 
 # Adjust our kernel's tunables
 echo 0 > $VM/dirty_writeback_centisecs
-echo 1 > $VM/swappiness
-echo 5 > $VM/vfs_cache_pressure
+echo 0 > $VM/swappiness
+echo 1 > $VM/vfs_cache_pressure
 echo 1 > $KR/sched_child_runs_first
+echo within_size > $THP/shmem_enabled
+echo always > $THP/defrag
+echo always > $THP/enabled
 
-# Allow swap to reach 97% before triggering LMKD
-resetprop -n ro.lmk.swap_free_low_percentage 3
+# Allow swap to reach 99% before triggering LMKD
+resetprop -n ro.lmk.swap_free_low_percentage 1
 
 # Disable SurfaceFlinger frame dropping, no but for real
-resetprop -d debug.sf.use_phase_offsets_as_durations
-resetprop -d debug.sf.late.sf.duration
-resetprop -d debug.sf.late.app.duration
-resetprop -d debug.sf.early.sf.duration
-resetprop -d debug.sf.early.app.duration
-resetprop -d debug.sf.earlyGl.sf.duration
-resetprop -d debug.sf.earlyGl.app.duration
-resetprop -d debug.sf.frame_rate_multiple_threshold
+resetprop -d debug.sf.use_phase_offsets_as_durations #1
+resetprop -d debug.sf.late.sf.duration               #10500000
+resetprop -d debug.sf.late.app.duration              #16600000
+resetprop -d debug.sf.early.sf.duration              #16600000
+resetprop -d debug.sf.early.app.duration             #16600000
+resetprop -d debug.sf.earlyGl.sf.duration            #16600000
+resetprop -d debug.sf.earlyGl.app.duration           #16600000
+resetprop -d debug.sf.frame_rate_multiple_threshold  #120
 
 # Raise the frequency of sampling regions in SurfaceFlinger
-## Defaults: 3000000, 200000000, 200000000
-resetprop -n debug.sf.region_sampling_duration_ns 4166666
-resetprop -n debug.sf.region_sampling_period_ns 99999984
-resetprop -n debug.sf.region_sampling_timer_timeout_ns 99999984
+resetprop -n debug.sf.region_sampling_duration_ns      8333333  #unset
+resetprop -n debug.sf.region_sampling_period_ns        99999984 #unset
+resetprop -n debug.sf.region_sampling_timer_timeout_ns 99999984 #unset
 
-# Restart SurfaceFlinger to take in the new values
-## Unfortunately increases boot time, need to set earlier in init before SF
-killall -9 surfaceflinger
+}
 
+zram() {
+  local gigs=${1:-1}
+
+  log "Dumping zram:"
+  zramcfg >> "$LOGFILE"
+
+  local kb=$(awk '/MemTotal/ {print $2}' /proc/meminfo)
+  local zramB=$(awk -v kb="$kb" -v g="$gigs" 'BEGIN{printf "%.0f\n", kb*1024 - g*1073741824}')
+  local sizeTxt=$(awk -v kb="$kb" 'BEGIN{print int(kb/1048576)+1}')
+  local zramTxt=$(awk -v b="$zramB" 'BEGIN{print int(b/1073741824)+1}')
+
+  log "Resizing ZRAM to ${zramTxt}/${sizeTxt}GB"
+  zramcfg -s "$zramB" >> "$LOGFILE"
 }
 
 #########
 
-case "$SHDIR" in
-  "$MAGISKPFDD")
-    echo "ptune: Waiting for boot complete"
-    while [ bootcomplete != "1" ]; do sleep 1; done
-    break;;
-  "$MAGISKSD")
-    echo "ptune: Initial boot"
-    break;;
-  *)
-    echo "ptune: Running unconditionally"
-esac
+logwipe
 
+log "Setting initial boot values"
 ptune
+
+# Avoid waiting to finalize values if we're already through init's boot sequence
+if [ "$(bootcomplete)" == "1" ]; then
+  zram
+
+  log "No need to finalize new values"
+  exit 0
+fi
+
+# Restart SurfaceFlinger to take in the new values
+## Unfortunately increases boot time, need to set earlier in init before SF
+log "Restarting SurfaceFlinger with force"
+killall -9 surfaceflinger
+
+log "Waiting for boot complete"
+while [ "$(bootcomplete)" != "1" ]; do sleep 1; done
+
+zram
+
+log "Finalizing values"
+ptune
+
+log "Pixel Tune is done!"
